Move calculations for suspicious metrics and main programming language to new files and change them to export functions #2710
